20230901-mediakit:  minor restructuring
