+++ PROJETOS/20230901-mediakit/mediakit_project/utils/__init__.py
@@ -17,10 +17,3 @@ def load_module(script_path, module_name="module"):
 
     return model_script
 
-
-class ModuleClass:
-    repo_dir = None
-
-    def __init__(self, **kwargs):
-        assert ModuleClass.repo_dir is not None
-        self.repo_dir = ModuleClass.repo_dir
